requirements.txt
anyio==4.2.0
Bottleneck==1.3.5
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cryptography==41.0.7
fastapi==0.103.0
h11==0.14.0
idna==3.4
numexpr==2.8.7
numpy==1.26.3
pandas==2.1.4
pip==23.3.1
pycparser==2.21
pydantic==1.10.12
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.20.0
wheel==0.41.2
